requirements_retrain.txt
pandas==2.2.3
xgboost==2.1.4
joblib==1.4.2
pymssql==2.3.2
python-dotenv==1.0.1
scikit-learn==1.3.0
xgboost==2.1.4